refactor: split hooks module into files (#4699)
* fix(hooks): link hooks page to swap trade state

* refactor: split hooks module into files
