+++ apps/cowswap-frontend/src/modules/hooks/const.ts
@@ -1,1 +1,2 @@
 export const HOOKS_TRAMPOLINE_ADDRESS = '0x01dcb88678aedd0c4cc9552b20f4718550250574'
+export const SBC_DEPOSIT_CONTRACT_ADDRESS = '0x0B98057eA310F4d31F2a452B414647007d1645d9'
