+++ apps/cowswap-frontend/src/modules/hooks/hooks/useHooks.ts
@@ -1,0 +1,7 @@
+import { useAtomValue } from 'jotai'
+
+import { hooksAtom } from '../state/hookDetailsAtom'
+
+export function useHooks() {
+  return useAtomValue(hooksAtom)
+}
