+++ apps/cowswap-frontend/src/modules/hooks/index.ts
@@ -1,3 +1,4 @@
-export * from './smart/hookButtons'
-export { useHooks } from './hooks'
 export { HookDappContext } from './context'
+export { PreHookButton } from './containers/PreHookButton'
+export { PostHookButton } from './containers/PostHookButton'
+export { useHooks } from './hooks/useHooks'
